Machine_learning/requirements.txt
Flask==3.1.3
mysql_connector_repackaged==0.3.1
numpy==2.4.4
python-dotenv==1.2.2
Requests==2.33.1
sounddevice==0.5.5
